Add patch fixing segfault on interrupted epoll_wait
The linux.c epoll(7) code did a naïve check on the return value of
epoll_wait(2) and proceeded to handle events even if none were
available (such as when epoll_wait was interrupted by SIGSTOP). This
resulted in a subsequent invalid pointer dereference and a segmentation
fault.

We fix this by importing a patch by Danny Guinther (thanks!).

LP: 1263808